fix(PR-4685): widen vote candidate/ballot indices from u8 to u32
Matt's CHANGES_REQUESTED: SubmitVoteCandidate::candidate_index was u8,
making candidates beyond index 255 unreachable or aliased in token-heavy
games (Council's Judgment / Prime Minister's Cabinet Room). Widened to u32
across all seven affected surfaces:

  - GameAction::SubmitVoteCandidate { candidate_index: u32 }
  - WaitingFor::VoteChoice.ballots: im::Vector<(PlayerId, u32)>
  - GameState::last_vote_ballots: im::Vector<(PlayerId, u32)>
  - VoteRoundState::ballots + append_vote_ballot_and_advance(idx: u32)
  - QuantityRef::VoteCount { choice_index: u32 }
  - PlayerFilter::VotedFor { choice_index: u32 }
  - resolve_tally / resolve_top_votes_tally ballot signatures
  - All as u8 casts in candidates.rs, engine_resolution_choices.rs,
    oracle_vote.rs, and vote.rs tests

Added regression test: object_vote_last_candidate_above_u8_max_is_selectable
creates 257 opponent permanents, votes unanimously for candidate 256 (>255),
and asserts it is exiled while un-voted permanents remain. Test fails with
the original u8 cast due to index wrap.

TieResolution::Breaker(u8) is intentionally left as u8 — it indexes named
vote option lists (bounded by Oracle text, never >255 choices in practice)
and is not involved in object-pool candidate dispatch.

Verification: cargo clippy -p engine -p phase-ai --all-targets -D warnings
clean; cargo test -p engine all pass including new regression test;
server-core clean.

+++ crates/engine/src/ai_support/candidates.rs
@@ -1294,7 +1294,7 @@ pub fn candidate_actions_broad(state: &GameState) -> Vec<CandidateAction> {
                     .map(|i| {
                         candidate(
                             GameAction::SubmitVoteCandidate {
-                                candidate_index: i as u8,
+                                candidate_index: i as u32,
                             },
                             TacticalClass::Selection,
                             Some(actor),
